requirements.txt
# Core ML/CV libraries
opencv-python
ultralytics
torch
torchvision

# Image processing
numpy
pillow

# Configuration
PyYAML

# Networking
requests

# Utilities
python-dotenv

# Logging
colorlog

# Notifications (optional)
plyer
